students/K33401/Savonik_Nikita/Lr3/requirements.txt
absl-py==0.9.0
addict==2.2.1
aioredis==1.3.1
aniso8601==8.0.0
appdirs==1.4.3
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.4
bottle==0.12.18
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cycler==0.10.0
Cython==0.29.20
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.1.2+cu102
Django==3.0.5
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.92.1
drf-yasg==1.20.0
entrypoints==0.3
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
fvcore==0.1.dev200325
gast==0.3.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
gym==0.17.2
h5py==2.10.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
imageio==2.8.0
imutils==0.5.3
incremental==17.5.0
inflection==0.5.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
livereload==2.6.3
lunr==0.5.8
lxml==4.5.2
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mat4py==0.4.3
mat73==0.45
matplotlib==3.2.1
metakernel==0.27.5
mistune==0.8.4
mkdocs==1.1.2
mkdocs-material==6.1.6
mkdocs-material-extensions==1.0.1
mock==4.0.2
mpmath==1.1.0
msgpack==1.0.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
oct2py==5.2.0
octave-kernel==0.32.0
opencv-python==4.2.0.32
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexif==0.15
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==1.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pycparser==2.20
pycuda==2019.1.2
pydot==1.4.1
pyglet==1.5.0
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylops==1.11.1
pymdown-extensions==8.0.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytools==2020.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
redis==3.5.3
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
sympy==1.6.2
tabulate==0.8.7
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.6
zope.interface==5.1.0
